Opa correção.

+++ src/main/java/br/com/eterniaserver/eterniaserver/generics/OnPlayerSignChange.java
@@ -18,7 +18,7 @@ public class OnPlayerSignChange implements Listener {
 
     @EventHandler (ignoreCancelled = true, priority = EventPriority.LOWEST)
     public void onPlayerSignChange(SignChangeEvent event) {
-        if (event.getPlayer().hasPermission("eternia.sign.color")) for (int i = 0; i < 4; i++) event.setLine(i, strings.getColor(event.getLine(i)));
+        if (event.getPlayer().hasPermission("eternia.sign.color")) for (byte i = 0; i < 4; i++) event.setLine(i, strings.getColor(event.getLine(i)));
     }
 
 }
